+++ CHANGELOG.md
@@ -2,10 +2,13 @@
 
 ## [Unreleased]
 
+MagicMirror² version >= 2.15.0 required.
+
 ### Fixed
 
 ### Added
 
+- Integrated MagicMirror logger on server side
 - Github actions `build` and `changelog`
 
 ### Changed